Disabled KeDelayExecutionThread, as it causes may delays. Disabled lots of high-level Xapi patches, as our kernel supports these functions correctly now.

+++ src/DxbxKrnl/uEmuKrnlNt.pas
@@ -2203,7 +2203,7 @@ function xboxkrnl_NtWaitForMultipleObjectsEx
   ): NTSTATUS; stdcall;
 // Branch:shogun  Revision:0.8.1-Pre2  Translator:PatrickvL  Done:100
 var
-  i: Integer;
+  i: ULONG;
 begin
   EmuSwapFS(fsWindows);
